refactor(python): share payment-channels core via _paycore
Move the reusable payment-channel machinery to pay_kit/_paycore/paymentchannels.py (mirrors the Go paycore layout) with a re-export shim at protocols/mpp/_paymentchannels.py so MPP and its tests are untouched. Fix the codama-py account decoders to skip the on-chain 1-byte account discriminator (codegen post-patch plus the committed channel/closedChannel accounts). Add SolanaRpc.get_account_info for on-chain channel-state reads.
