+++ skills/pay-sdk-implementation/codegen/generate-payment-channels-client-py.ts
@@ -2,7 +2,7 @@
  * Generate the pay-kit Python payment-channels client from the upstream
  * `Moonsong-Labs/solana-payment-channels` Codama IDL.
  *
- * Mirrors generate-payment-channels-client.ts (Rust) — both scripts read the
+ * Mirrors generate-payment-channels-client.ts (Rust) - both scripts read the
  * vendored IDL at `<repo-root>/idl/payment-channels.json`. This one renders a
  * Python client into `python/src/pay_kit/protocols/programs/paymentchannels/`
  * using the community `codama-py` renderer (Solana-ZH/codama-py).
@@ -10,8 +10,8 @@
  * codama-py cannot be consumed as an npm/git dependency yet: its package.json
  * ships only `dist` (not committed, and its build is currently broken
  * upstream), so a git install packs an empty module. Until a fixed release
- * exists this script instead clones the repo at a pinned commit — the merge
- * of Solana-ZH/codama-py#10, which fixed PDA seed rendering — and drives its
+ * exists this script instead clones the repo at a pinned commit - the merge
+ * of Solana-ZH/codama-py#10, which fixed PDA seed rendering - and drives its
  * own `genpy` CLI, exactly as the upstream README documents.
  *
  * Output:
@@ -27,7 +27,7 @@ const CODAMA_PY_REPO = 'https://github.com/Solana-ZH/codama-py.git';
 const CODAMA_PY_COMMIT = 'fcc75fc8abdf18cf4e0b3e4ae9338ddb60deb2e1';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
-// Script lives at skills/pay-sdk-implementation/codegen/ — climb three
+// Script lives at skills/pay-sdk-implementation/codegen/ - climb three
 // levels to land at the repository root.
 const repoRoot = path.resolve(__dirname, '..', '..', '..');
 
@@ -71,6 +71,46 @@ const sentinel = path.join(pyClientDir, 'program_id.py');
 if (!fs.existsSync(sentinel)) {
     console.error(`[codegen] Render produced no output at ${path.relative(repoRoot, pyClientDir)}`);
     process.exit(1);
+}
+
+// Post-generation patch: restore the account discriminator skip in `decode`.
+//
+// For ACCOUNT types the codama IDL keeps a leading 1-byte account
+// discriminator as field 0 (a numberTypeNode/u8), and the official Go and
+// Rust renderers decode it first. codama-py, however, DROPS that field from
+// the generated Borsh layout and reads `<Cls>.layout.parse(data)` from offset
+// 0, so every subsequent field is misaligned by one byte against real
+// on-chain account data. We patch each generated account decoder to parse
+// from offset 1 (`<Cls>.layout.parse(data[1:])`), skipping the discriminator
+// byte the layout omits. This keeps the change deterministic and idempotent,
+// and faithful to the IDL the Go/Rust clients honour.
+const accountsDir = path.join(pyClientDir, 'accounts');
+if (fs.existsSync(accountsDir)) {
+    const accountFiles = fs
+        .readdirSync(accountsDir)
+        .filter((name) => name.endsWith('.py') && name !== '__init__.py');
+    for (const name of accountFiles) {
+        const filePath = path.join(accountsDir, name);
+        const original = fs.readFileSync(filePath, 'utf8');
+        // Match the generated decode body: `dec = <Cls>.layout.parse(data)`.
+        const pattern = /(\bdec\s*=\s*\w+\.layout\.parse\()data(\))/;
+        const alreadyPatched = /\bdec\s*=\s*\w+\.layout\.parse\(data\[1:\]\)/.test(original);
+        if (alreadyPatched) {
+            continue; // idempotent: re-running codegen must not double-patch.
+        }
+        if (!pattern.test(original)) {
+            console.error(
+                `[codegen] Expected '<Cls>.layout.parse(data)' in ${path.relative(repoRoot, filePath)} but did not find it.`,
+            );
+            console.error(
+                `[codegen] codama-py may have changed its account decoder; the discriminator-skip patch cannot be applied safely.`,
+            );
+            process.exit(1);
+        }
+        const patched = original.replace(pattern, '$1data[1:]$2');
+        fs.writeFileSync(filePath, patched);
+        console.log(`[codegen]   patched discriminator skip in accounts/${name}`);
+    }
 }
 
 console.log(`[codegen] Done.`);
